removed all unneeded code for sbus test
